% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioE.R
\name{bioE}
\alias{bioE}
\title{Fish Bioenergetics model}
\usage{
bioE(par, data = list(W, TempC, Eprey, Epred, indgst, diet, fTCmodel = NA,
  fTrmodel = NA, velmodel = NA))
}
\arguments{
\item{par}{is a list of parameters used for the model}

\item{data}{is a list with a daily observation of weight, temperature, prey energy density, and predator energy density}

\item{data$W}{weight of the fish in grams}
}
\description{
Rcode developed by: Kirstin Holsman
kirstin.holsman@noaa.gov
This function calculates the weight and tempature specific physiological rates given a set of parameters and data inputs. Predicts consumption or growth from a fit or set pvalue using model.type
The model below is based on the "Wisconsin" Fish Bioenergetics model (Hanson et al. 1997)
model.type =
1) fit to observed growth by adjusting P-value
2) predict growth from set p-value (in par)
}
\examples{
plk_par<-data.frame(RFR=1, Qox=13560,Ceq=2,Req=2,Weq=1,Tco=10,Tcm=15,QC=2.6,CA=0.119,CB=-0.46, 
                    RA=0.0075,RB=-0.251, QR=2.6,Tro=13,Trm=18,SA=0.125, Am=2,FA=0.15,UA=0.11)
ebs_data<-list(W=100,TempC=0:10,Eprey=5539.6,Epred=4184,indgst=0,diet=0)
bioE(data=ebs_data,par=plk_par)
fTfun()
}
\keyword{Temperature,}
\keyword{consumption}
\keyword{scaling,}
