% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Halibut.R
\name{fish_nrg}
\alias{fish_nrg}
\title{Halibut BIOE model}
\usage{
fish_nrg(par, data)
}
\arguments{
\item{par}{is a list of parameters used for the model}

\item{data}{is a list with a daily observation of weight, temperature, prey energy density, and predator energy density}

\item{data$W}{weight of the fish in grams}
}
\description{
This function calculates the weight and tempature specific physiological paramaters given a set of parameters and data inputs. Predicts consumption or growth from a fit or set pvalue using model.type
model.type =
1) fit to observed growth by adjusting P-value
2) predict growth from set p-value (in par)
}
\examples{
par<-c(2)
  pars<-data.frame(CA=Cparm2[1],CB=Cparm2[2],Tco=exp(Ft_model$par)[1],Tcm=18,Qc=exp(Ft_model$par)[2],sigma=exp(Ft_model$par)[3])
pars2<-data.frame(CA=Cparm[1],CB=Cparm[2],Tco=exp(Ft_model$par)[1],Tcm=18,Qc=exp(Ft_model$par)[2],sigma=exp(Ft_model$par)[3])
	pars3<-data.frame(CA=Cparm[1],CB=Cparm[2],Tco=exp(Ft_model$par)[1],Tcm=18,CQ=exp(Ft_model$par)[2],O2cal=13560)
Halibut_parms<-list(first_choice=newPAR2,second_choice=Cparm2)
	PARMS<-c(pars,Rpar,UA=UA,UA.sigma=UA.sigma,FA=FA,SA=SA, SA.sigma=SA.sigma)
PARMS_USE<-c(pars2,Rpar,UA=UA,UA.sigma=UA.sigma,FA=FA,SA=SA, SA.sigma=SA.sigma)
fish_nrg(par=PARMS_USE,data=list(TempC=10,W=100,Eprey=4000,Pvalue=1))
}
\keyword{Temperature,}
\keyword{consumption}
\keyword{scaling,}
