% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioE.R
\name{fTC_fun}
\alias{fTC_fun}
\title{Consumption function; Fish Bioenergetics model
This function calculates the temperature and weight specific consumption rate for an individual fish}
\usage{
fTC_fun(par, data)
}
\arguments{
\item{par}{is a list of parameters used for the model}

\item{data}{is a list with a daily observation of weight, temperature, prey energy density, and predator energy density}

\item{Ceq}{is a par list object (par$Ceq) that specifies the type of consumption equation to use: 0 = specify function using data$fTCmodel, 1= exponential (type 1 from Fish Bioenergetics), 2 = type 2, 3 = type 3}
}
\description{
Rcode developed by: Kirstin Holsman
kirstin.holsman@noaa.gov
This function calculates the weight and tempature specific physiological paramaters given a set of parameters and data inputs.
The model below is based on the "Wisconsin" Fish Bioenergetics model (Hanson et al. 1997)
}
\examples{
c_data<-ebs_data
c_data$fTCmodel<-function(TempC){-.5*TempC}
fTC_fun(par=plk_par,data=c_data)
}
\keyword{Respiration}
\keyword{Temperature,}
\keyword{scaling,}
