% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrg.R
\docType{package}
\name{NRG}
\alias{NRG}
\alias{NRG-package}
\title{NRG: A package for simulating growth using bioenergetics models}
\description{
The NRG package contains functions to predict consumption,
respiration, waste, and growth of fish based on bioenergetic
energy balance equations.
}
\section{NRG functions}{

The NRG functions ...
}

