% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Halibut.R
\name{nrg_sim}
\alias{nrg_sim}
\title{simple simulation model}
\usage{
nrg_sim(Tdat, Pvalue, preyEnergy, fishEnergy = 20050, W0)
}
\arguments{
\item{par}{is a list of parameters used for the model}

\item{data}{is a list with a daily observation of weight, temperature, prey energy density, and predator energy density}

\item{data$W}{weight of the fish in grams}
}
\description{
This function calculates the weight and tempature specific physiological paramaters given a set of parameters and data inputs. Predicts consumption or growth from a fit or set pvalue using model.type
model.type =
1) fit to observed growth by adjusting P-value
2) predict growth from set p-value (in par)
}
\examples{
nrg_sim()
}
\keyword{Temperature,}
\keyword{consumption}
\keyword{scaling,}
