% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deb.R
\docType{data}
\name{par}
\alias{par}
\title{Fish Dynamic Energy Budget Model}
\format{An object of class \code{numeric} of length 16.}
\usage{
par
}
\arguments{
\item{par}{is a list of parameters used for the model}

\item{input}{is a list with inputs for the model including a daily observation of weight, temperature, prey energy density, and predator energy density}

\item{k}{Fraction of energy to somatic growth and maintenance}

\item{v}{Energy conductance (cm d-1), temperature dependent}

\item{JEAm}{Maximum assimilation rate per surface area (mmol d-1 cm-2)}

\item{kj}{Fraction of maturity maintenancec (d-1), temperature dependent}

\item{EpH}{Maturity energy at puberty (kJ)}

\item{gamma}{Growth rate of roe (d-1)}

\item{TA}{Arrhenius temperature (K)}

\item{Tr}{Reference temperature (K)}

\item{Mv}{Volume specific structural mass (mmol cm-3)}

\item{muE}{Chemical potential (J mmol-3)}

\item{s}{Shape coefficient that relates volumetric length to actual length; often obtained by fitting a weight–length relationship of the type M=(s*length)^3}

\item{Lm}{Maximum structural (volumetric) length (cm)}

\item{dv}{Density of structural volume (g cm-3)}

\item{pE}{Energy reserve density (kJ g-1)}

\item{pR}{Energy density of roe (kJ g-1)}
}
\description{
Rcode developed by: Kirstin Holsman  
kirstin.holsman@noaa.gov
This function calculates the dynamic energy budget as outlined by B. Einarsson et al. / Journal of Theoretical Biology 281 (2011) 1–8
According to DEB theory, each individual allocates a fixed
fraction k of utilized energy from reserves to growth and somatic
maintenance. The rest, (1-k), is then allocated to maturity maintenance and reproduction.
}
\examples{
plk_par<-data.frame(RFR=1, Qox=13560,Ceq=2,Req=2,Weq=1,Tco=10,Tcm=15,QC=2.6,CA=0.119,CB=-0.46, 
                    RA=0.0075,RB=-0.251, QR=2.6,Tro=13,Trm=18,SA=0.125, Am=2,FA=0.15,UA=0.11)
ebs_data<-list(W=100,TempC=0:10,Eprey=5539.6,Epred=4184,indgst=0,diet=0,fTCmodel=NA,fTrmodel=NA,velmodel=NA)
bioE(data=ebs_data,par=plk_par)
fTfun() 
}
\keyword{Temperature,}
\keyword{consumption}
\keyword{scaling,}
