% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioE.R
\name{Waste_fun}
\alias{Waste_fun}
\title{Waste function; Fish Bioenergetics model
This function calculates the temperature and weight specific estimates of lost energey (waste) for an individual fish}
\usage{
Waste_fun(par, data)
}
\arguments{
\item{par}{is a list of parameters used for the model}

\item{data}{is a list with a daily observation of weight, temperature, prey energy density, and predator energy density}

\item{data$W}{weight of the fish in grams}
}
\description{
Rcode developed by: Kirstin Holsman
kirstin.holsman@noaa.gov
This function calculates the weight and tempature specific physiological paramaters given a set of parameters and data inputs. 
The model below is based on the "Wisconsin" Fish Bioenergetics model (Hanson et al. 1997)
}
\examples{
Waste_fun()
}
\keyword{Respiration}
\keyword{Temperature,}
\keyword{scaling,}
