% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioE.R
\name{bioE}
\alias{bioE}
\title{Fish Bioenergetics model}
\usage{
bioE(par, data = list(W, TempC, Eprey, Epred, indgst, diet, fTCmodel = NA,
  fTrmodel = NA, velmodel = NA))
}
\arguments{
\item{par}{is a list of parameters used for the model}

\item{data}{is a list with inputs for the model including a daily observation of weight, temperature, prey energy density, and predator energy density}

\item{W}{weight of the fish in grams}

\item{TempC}{is the temperature the fish experiences (in deg C)}

\item{Eprey}{is the energey density of the prey in Joules per gram of prey}

\item{Epred}{is the energey density of the fish predator in Joules per gram of pred}

\item{indgst}{is the indigestable proportion of each prey item in the diet}

\item{diet}{is the proportion by weight of each prey item in the diet}

\item{fTCmodel}{is the optional user specified function for the temperature scaling component of C; set to NA by default}

\item{fTrmodel}{the optional user specified function for the temperature scaling component of R; set to NA by default}

\item{velmodel}{the optional user specified function for the temperature scaling component of swim velocity; set to NA by default}
}
\description{
Rcode developed by: Kirstin Holsman
kirstin.holsman@noaa.gov
Based on the Fish Bioenergetics Model 3.0, see update 4.0 by Deslauriers, D., 
Chipps, S.R., Breck, J.E., Rice, J.A. & Madenjian, C.P. 2017. Fish Bioenergetics 4.0: 
An R- based modeling application. Fisheries, 42(11): 586–596.
}
\details{
This function calculates the weight and tempature specific physiological rates given a set of parameters and data inputs. Predicts consumption or growth from a fit or set pvalue using model.type
The model below is based on the "Wisconsin" Fish Bioenergetics model (Hanson et al. 1997)
}
\examples{
plk_par<-data.frame(RFR=1, Qox=13560,Ceq=2,Req=2,Weq=1,Tco=10,Tcm=15,QC=2.6,CA=0.119,CB=-0.46, 
                    RA=0.0075,RB=-0.251, QR=2.6,Tro=13,Trm=18,SA=0.125, Am=2,FA=0.15,UA=0.11)
ebs_data<-list(W=100,TempC=0:10,Eprey=5539.6,Epred=4184,indgst=0,diet=0,fTCmodel=NA,fTrmodel=NA,velmodel=NA)
bioE(data=ebs_data,par=plk_par)
fTfun()
13.56 J mg-1 Brett & Groves 1979 Qox for Fish 
}
\keyword{Temperature,}
\keyword{consumption}
\keyword{scaling,}
